Move FourSubstationsNodeBreakerWithExtensionsFactory to iidm-test
