+++ iidm/iidm-extensions/pom.xml
@@ -54,11 +54,6 @@
             <artifactId>powsybl-iidm-api</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>powsybl-iidm-test</artifactId>
-            <version>${project.version}</version>
-        </dependency>
 
         <!-- test dependencies -->
         <dependency>
